Factors out processInput method in Duke to its own parser class modify duke to make use of new parser class
Duke main used to have own process input method, cluttering the code

Abstracts methods related to processing input into its own Parser class
