+++ src/main/java/Duke.java
@@ -4,12 +4,13 @@ import java.util.Scanner;
 
 import storage.Storage;
 import ui.Ui;
+import parser.Parser;
 import exceptions.*;
 import commands.*;
 import tasks.*;
 
 public class Duke {
-	private static ArrayList<Task> TASKLIST = new ArrayList<>();
+//	private static ArrayList<Task> TASKLIST = new ArrayList<>();
 	private TaskList tasks;
 	private static Storage storage;
 	private static Ui ui;
@@ -35,23 +36,18 @@ public class Duke {
 		while (true) {
 			try {
 				String input = sc.nextLine();
-				cmd = processInput(input);
 				ui.printLine();
+				cmd = Parser.processInput(input);
 				cmd.execute(tasks, ui, storage);
-				ui.printLine();
+				tasks = cmd.getList();
 				if (cmd.ends()) {
 					break;
 				}
 			} catch (DukeException e) {
 				e.printError();
+			} finally {
+				ui.printLine();
 			}
-		}
-
-		try{
-			storage.saveFile(tasks);
-		} catch (IOException e) {
-			System.out.println("An error occurred");
-			e.printStackTrace();
 		}
 	}
 
@@ -59,62 +55,62 @@ public class Duke {
 		new Duke("data/tasks.txt").run();
 	}
 
-	private static Command processInput(String input) throws DukeException {
-		Command cmd = null;
-		String[] inputSplit = input.split(" ", 2);
-		String commandType = inputSplit[0].toLowerCase();
-		switch (commandType) {
-		case "bye":
-			cmd = new ByeCommand();
-			break;
-		case "list":
-			cmd = new ListCommand(TASKLIST);
-			break;
-		case "mark":
-		case "unmark":
-		case "delete":
-			String detail = inputSplit[1];
-			if (!isInteger(detail)) {
-				throw new IncorrectValueException();
-			}
-			int inputVal = Integer.parseInt(detail);
-			if (inputVal > TASKLIST.size()) {
-				throw new OutOfRangeException();
-			}
-			switch (commandType) {
-			case "mark":
-				cmd = new MarkCommand(TASKLIST, inputVal);
-				break;
-			case "unmark":
-				cmd = new UnMarkCommand(TASKLIST, inputVal);
-				break;
-			case "delete":
-				cmd = new DeleteCommand(TASKLIST, inputVal);
-				TASKLIST = cmd.getList();
-				break;
-			}
-			break;
-		case "todo":
-		case "event":
-		case "deadline":
-			if (inputSplit.length < 2) {
-				throw new EmptyDescriptionException(commandType);
-			}
-			cmd = new AddCommand(TASKLIST, commandType, inputSplit[1]);
-			TASKLIST = cmd.getList();
-			break;
-		default:
-			throw new UnknownCommandException();
-		}
-		return cmd;
-	}
-
-	private static boolean isInteger(String input) {
-		for (int i = 0; i < input.length(); i++) {
-			if (!Character.isDigit(input.charAt(i))) {
-				return false;
-			}
-		}
-		return true;
-	}
+//	private static Command processInput(String input) throws DukeException {
+//		Command cmd = null;
+//		String[] inputSplit = input.split(" ", 2);
+//		String commandType = inputSplit[0].toLowerCase();
+//		switch (commandType) {
+//		case "bye":
+//			cmd = new ByeCommand();
+//			break;
+//		case "list":
+//			cmd = new ListCommand(TASKLIST);
+//			break;
+//		case "mark":
+//		case "unmark":
+//		case "delete":
+//			String detail = inputSplit[1];
+//			if (!isInteger(detail)) {
+//				throw new IncorrectValueException();
+//			}
+//			int inputVal = Integer.parseInt(detail);
+//			if (inputVal > TASKLIST.size()) {
+//				throw new OutOfRangeException();
+//			}
+//			switch (commandType) {
+//			case "mark":
+//				cmd = new MarkCommand(TASKLIST, inputVal);
+//				break;
+//			case "unmark":
+//				cmd = new UnMarkCommand(TASKLIST, inputVal);
+//				break;
+//			case "delete":
+//				cmd = new DeleteCommand(TASKLIST, inputVal);
+//				TASKLIST = cmd.getList();
+//				break;
+//			}
+//			break;
+//		case "todo":
+//		case "event":
+//		case "deadline":
+//			if (inputSplit.length < 2) {
+//				throw new EmptyDescriptionException(commandType);
+//			}
+//			cmd = new AddCommand(TASKLIST, commandType, inputSplit[1]);
+//			TASKLIST = cmd.getList();
+//			break;
+//		default:
+//			throw new UnknownCommandException();
+//		}
+//		return cmd;
+//	}
+//
+//	private static boolean isInteger(String input) {
+//		for (int i = 0; i < input.length(); i++) {
+//			if (!Character.isDigit(input.charAt(i))) {
+//				return false;
+//			}
+//		}
+//		return true;
+//	}
 }
